working with our final data
this app works with sleep vs cortisol data looking for interesting correlations. also adds a new column to the display table and populates the column with new values for change in cortisol calculated from existing data.
next I plan to incorporate sophia's code for the line of best fit checkbox.